Context methods only for go1.7+
